in game menu functions refactor
